+++ Sources/Controllers/PlaybackController.m
@@ -42,7 +42,7 @@ BOOL playOnStart = YES;
 - (void) awakeFromNib {
   NSNotificationCenter *center = [NSNotificationCenter defaultCenter];
 
-  NSWindow *window = [[NSApp delegate] window];
+  NSWindow *window = [HMSAppDelegate window];
   [center addObserver:self
              selector:@selector(stopUpdatingProgress)
                  name:NSWindowWillCloseNotification
@@ -96,14 +96,6 @@ BOOL playOnStart = YES;
      name:StationDidPlaySongNotification
      object:nil];
 
-  CIFilter *volumeSliderFilter = [CIFilter filterWithName:@"CIPhotoEffectMono"];
-  if (volumeSliderFilter != nil)
-    [volume setContentFilters:@[volumeSliderFilter]];
-
-  CIFilter *playbackProgressFilter = [CIFilter filterWithName:@"CIPhotoEffectMono"];
-  if (playbackProgressFilter != nil)
-    [playbackProgress setContentFilters:@[playbackProgressFilter]];
-  
   // NSDistributedNotificationCenter is for interprocess communication.
   [[NSDistributedNotificationCenter defaultCenter] addObserver:self
                                                       selector:@selector(pauseOnScreensaverStart:)
@@ -190,18 +182,18 @@ BOOL playOnStart = YES;
 }
 
 - (Pandora*) pandora {
-  return [[NSApp delegate] pandora];
+  return [HMSAppDelegate pandora];
 }
 
 - (void) reset {
   [self playStation:nil];
 
-  NSString *path = [[NSApp delegate] stateDirectory:@"station.savestate"];
+  NSString *path = [HMSAppDelegate stateDirectory:@"station.savestate"];
   [[NSFileManager defaultManager] removeItemAtPath:path error:nil];
 }
 
 - (void) show {
-  [[NSApp delegate] setCurrentView:playbackView];
+  [HMSAppDelegate setCurrentView:playbackView];
 }
 
 - (void) showSpinner {
@@ -215,7 +207,7 @@ BOOL playOnStart = YES;
 }
 
 - (BOOL) saveState {
-  NSString *path = [[NSApp delegate] stateDirectory:@"station.savestate"];
+  NSString *path = [HMSAppDelegate stateDirectory:@"station.savestate"];
   if (path == nil) {
     return NO;
   }
@@ -268,7 +260,7 @@ BOOL playOnStart = YES;
     return;
 
   QLPreviewPanel *previewPanel = [QLPreviewPanel sharedPreviewPanel];
-  if (previewPanel.currentController != [NSApp delegate])
+  if (previewPanel.currentController != HMSAppDelegate)
     return;
 
   if (hasArt)
@@ -326,7 +318,7 @@ BOOL playOnStart = YES;
     NSLogd(@"Skipping loading image");
   }
 
-  [[NSApp delegate] setCurrentView:playbackView];
+  [HMSAppDelegate setCurrentView:playbackView];
 
   [songLabel setStringValue: [song title]];
   [songLabel setToolTip:[song title]];
@@ -347,7 +339,7 @@ BOOL playOnStart = YES;
     [toolbar setSelectedItemIdentifier:nil];
   }
 
-  [[[NSApp delegate] history] addSong:song];
+  [[HMSAppDelegate history] addSong:song];
   [self hideSpinner];
 }
 
@@ -373,7 +365,7 @@ BOOL playOnStart = YES;
   [[NSUserDefaults standardUserDefaults] setObject:[station stationId]
                                             forKey:LAST_STATION_KEY];
   
-  [[NSApp delegate] showLoader];
+  [HMSAppDelegate showLoader];
 
   if (playOnStart) {
     [station play];
@@ -433,8 +425,8 @@ BOOL playOnStart = YES;
     }
   }
 
-  if ([[[NSApp delegate] history] selectedItem] == song) {
-    [[[NSApp delegate] history] updateUI];
+  if ([[HMSAppDelegate history] selectedItem] == song) {
+    [[HMSAppDelegate history] updateUI];
   }
 }
 
@@ -489,7 +481,7 @@ BOOL playOnStart = YES;
 /* Load more songs manually */
 - (IBAction)loadMore: (id)sender {
   [self showSpinner];
-  [[NSApp delegate] setCurrentView:playbackView];
+  [HMSAppDelegate setCurrentView:playbackView];
 
   if ([playing playingSong] != nil) {
     [playing retry];
@@ -683,7 +675,7 @@ BOOL playOnStart = YES;
 
 - (NSRect)previewPanel:(QLPreviewPanel *)panel sourceFrameOnScreenForPreviewItem:(id <QLPreviewItem>)item {
   NSRect frame = [art frame];
-  frame = [[[NSApp delegate] window] convertRectToScreen:frame];
+  frame = [[HMSAppDelegate window] convertRectToScreen:frame];
 
   frame = NSInsetRect(frame, 1, 1); // image doesn't extend into the button border