+++ Sources/Integration/AppleScript.m
@@ -14,21 +14,21 @@ int savedVolume = 0;
 
 @implementation PlayCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   return @([playback play]);
 }
 @end
 
 @implementation PauseCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   return @([playback pause]);
 }
 @end
 
 @implementation PlayPauseCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [playback playpause:self];
   return self;
 }
@@ -36,28 +36,28 @@ int savedVolume = 0;
 
 @implementation SkipCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [playback next:self];
   return self;
 }
 @end
 @implementation ThumbsUpCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [playback like:self];
   return self;
 }
 @end
 @implementation ThumbsDownCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [playback dislike:self];
   return self;
 }
 @end
 @implementation RaiseVolumeCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   int volume = [playback getIntVolume];
   [playback setIntVolume:volume + 7];
   NSLogd(@"Raised volume to: %d", [playback getIntVolume]);
@@ -66,7 +66,7 @@ int savedVolume = 0;
 @end
 @implementation LowerVolumeCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   int volume = [playback getIntVolume];
   [playback setIntVolume:volume - 7];
   NSLogd(@"Lowered volume to: %d", [playback getIntVolume]);
@@ -75,7 +75,7 @@ int savedVolume = 0;
 @end
 @implementation FullVolumeCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [playback setIntVolume:100];
   NSLogd(@"Changed volume to: %d", [playback getIntVolume]);
   return self;
@@ -83,7 +83,7 @@ int savedVolume = 0;
 @end
 @implementation MuteCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   savedVolume = [playback getIntVolume];
   [playback setIntVolume:0];
   NSLogd(@"Changed volume to: %d", [playback getIntVolume]);
@@ -92,7 +92,7 @@ int savedVolume = 0;
 @end
 @implementation UnmuteCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [playback setIntVolume:savedVolume];
   NSLogd(@"Changed volume to: %d", [playback getIntVolume]);
   return self;
@@ -100,7 +100,7 @@ int savedVolume = 0;
 @end
 @implementation TiredCommand
 - (id) performDefaultImplementation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [playback tired:self];
   return self;
 }
@@ -109,17 +109,17 @@ int savedVolume = 0;
 @implementation NSApplication (HermesScripting)
 
 - (NSNumber*) volume {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   return @([playback getIntVolume]);
 }
 
 - (void) setVolume: (NSNumber*) vol {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [playback setIntVolume:[vol intValue]];
 }
 
 - (int) playbackState {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   Station *playing = [playback playing];
   if (playing == nil) {
     return PlaybackStateStopped;
@@ -131,20 +131,20 @@ int savedVolume = 0;
 
 - (NSNumber *) playbackPosition {
   double progress;
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [[playback playing] progress:&progress];
   return @(progress);
 }
 
 - (NSNumber *) currentSongDuration {
   double duration;
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   [[playback playing] duration:&duration];
   return @(duration);
 }
 
 - (void) setPlaybackState: (int) state {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   switch (state) {
     case PlaybackStateStopped:
     case PlaybackStatePaused:
@@ -161,12 +161,12 @@ int savedVolume = 0;
 }
 
 - (Station*) currentStation {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   return [playback playing];
 }
 
 - (void) setCurrentStation:(Station *)station {
-  HermesAppDelegate *delegate = [NSApp delegate];
+  HermesAppDelegate *delegate = HMSAppDelegate;
   PlaybackController *playback = [delegate playback];
   [playback playStation:station];
   StationsController *stations = [delegate stations];
@@ -174,12 +174,12 @@ int savedVolume = 0;
 }
 
 - (NSArray*) stations {
-  HermesAppDelegate *delegate = [NSApp delegate];
+  HermesAppDelegate *delegate = HMSAppDelegate;
   return [[delegate pandora] stations];
 }
 
 - (Song*) currentSong {
-  PlaybackController *playback = [[NSApp delegate] playback];
+  PlaybackController *playback = [HMSAppDelegate playback];
   return [[playback playing] playingSong];
 }