+++ Sources/Views/HistoryCollectionView.m
@@ -14,7 +14,7 @@
 
 - (void)keyDown:(NSEvent *)theEvent {
   if ([[theEvent characters] isEqualToString:@" "]) {
-    [[[NSApp delegate] playback] playpause:nil];
+    [[HMSAppDelegate playback] playpause:nil];
   } else {
     [super keyDown:theEvent];
   }