+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * [added by @paullj1] Allow skipping songs (if using banners) and liking/disliking songs (if using alerts) from OS X notifications (#273)
 * [changed by @paullj1] Make Hermes OS X song notifications look more like iTunes notifications (#273)
+* [changed by @nriley] Simplify and reduce screen space used by main window
+* [fixed by @nriley] Work around OS X 10.10 bug causing strange drawer background coloring
+* [fixed by @nriley] Work around macOS 10.12 bug which caused volume slider and song progress not to display
 
 # Version 1.2.8 (6/24/16)